Calibration + Merge
So this version consists of all of the files merged from our two
"branches" of work. There is also working debounced calibration in
main.c (feel free to move it anywhere you need it) It also fixes a
number of bugs (like array overflow) that showed up only after merging
the two files.